Upload the app with the basic functions for Create a Flashcard and popup
