+++ app/assets/stylesheets/pages/_home.scss
@@ -79,6 +79,7 @@
   text-align: center;
   padding: 10px;
   box-shadow: 6px 6px 1px 1px rgba(51, 51, 52, 0.1);
+  text-decoration: none;
 }
 
 // .home-btn:hover {
