+++ app/assets/stylesheets/pages/_user_flashcards.scss
@@ -1,30 +1,146 @@
+.flash-background {
+  background-image: url("background_pic1.png");
+  background-repeat: no-repeat;
+  background-size: 100% auto;
+  background-position: calc(100% - 5px) calc(100% - 50px);
+  min-height: 100vh;
+  z-index: 0;
+  margin-top: 5px;
+}
+
 .autocomplete {
   position: relative;
-  display: inline-block;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  background-color: #E7EDFF;
+  padding: 25px 0px;
+  border-radius: 25px;
+  box-shadow: 3px 3px 1px 1px rgba(51, 51, 52, 0.1);
 }
 #search-box {
-  background-color: red;
   width: 300px;
   padding: 10px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
+  border: none;
+  border-radius: 25px;
+  background-color: #fff;
+  box-shadow: 3px 3px 1px 1px rgba(51, 51, 52, 0.1);
 }
 #suggestions {
   list-style-type: none;
-  padding: 0;
-  position: absolute;
+  padding: 5px 0px;
+  position: relative;
   background-color: #fff;
-  border: 1px solid #ccc;
-  border-top: none;
+  border: none;
   display: none;
   max-height: 150px;
   overflow-y: auto;
   width: 298px;
+  border-radius: 25px;
+  z-index: 2;
 }
 #suggestions li {
   padding: 10px;
-  cursor: pointer;
 }
 #suggestions li:hover {
   background-color: #F1F1F1;
+}
+
+.Text-Flash {
+  font-size: 18px;
+  font-weight: bold;
+  margin-left: 5px;
+}
+
+.flashcard_word {
+  background-color: #E7EDFF;
+  padding: 25px 0px;
+  border-radius: 25px;
+  margin-bottom: 30px;
+  display: flex;
+  flex-direction: column;
+  box-shadow: 3px 3px 1px 1px rgba(51, 51, 52, 0.1);
+}
+
+.flashcard_word input {
+  box-shadow: 3px 3px 1px 1px rgba(51, 51, 52, 0.1);
+}
+
+.flashcard_word h5 {
+  vertical-align: bottom;
+  display: inline;
+}
+
+.flashcard_word input,
+.flashcard_word h5 {
+  padding: 10px;
+  margin: 5px 20px;
+}
+
+.flashcard_word .line {
+  height: 2px;
+  background-color: #adc0f9;
+  width: 100%;
+}
+
+#word-flash {
+  width: 300px;
+  padding: 10px;
+  border: none;
+  border-radius: 25px;
+  background-color: #fff;
+  margin-bottom: 20px;
+}
+
+.btn-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 40px;
+}
+
+.flash-btn {
+  font-size: 20px;
+  padding: 10px 40px;
+}
+
+.popup {
+  display: none;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  z-index: 9999;
+  background-color: #FCA401;
+  border-radius: 25px;
+  min-height: 50vh;
+  box-shadow: 500px 500px 500px 500px rgba(51, 51, 52, 1);
+}
+
+.popup-content {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding: 100px;
+}
+
+.popup-btn {
+  background-color: #5b4cff;
+  color: white;
+  font-size: 20px;
+  border-radius: 50px;
+  border: none;
+  margin: 5px;
+  text-align: center;
+  padding: 10px;
+  width: 150px;
+  box-shadow: 6px 6px 1px 1px rgba(51, 51, 52, 0.1);
+  text-decoration: none;
+}
+
+.popupText {
+  font-size: 50px;
+  font-weight: bold;
 }
